First marching cubes drawing. Looks ridiculous --> no normals.

+++ src/GridCell.cpp
@@ -1,6 +1,6 @@
 #include "GridCell.h"
 
-GridCell::GridCell(const glm::vec3 &position, Particle* value[8]) :
+GridCell::GridCell(const glm::ivec3 &position, Particle* value[8]) :
         position(position) {
     for(auto i = 0; i < 8; i++) {
         this->value[i] = value[i];
