+++ src/GridCell.h
@@ -7,7 +7,7 @@ class Particle;
 class GridCell {
 public:
     GridCell();
-    GridCell(const glm::vec3 &position, Particle* value[8]);
+    GridCell(const glm::ivec3 &position, Particle* value[8]);
 
     glm::vec3 position;
     Particle* value[8];
